paper insertion
