+++ Content.Server/_CE/Science/CEResearchTableSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared._CE.Science;
+
+namespace Content.Server._CE.Science;
+
+public sealed partial class CEResearchTableSystem : CESharedResearchTableSystem;
