+++ Content.Shared/_CE/Science/Components/CEResearchTableComponent.cs
@@ -1,3 +1,5 @@
+using Robust.Shared.Serialization;
+
 namespace Content.Shared._CE.Science.Components;
 
 /// <summary>
@@ -8,4 +10,10 @@ public sealed partial class CEResearchTableComponent : Component
 {
     [DataField]
     public string PaperSlotId = "paper";
+}
+
+[Serializable, NetSerializable]
+public enum CEResearchTableVisuals : byte
+{
+    HasPaper,
 }
