--- Resources/Textures/_CE/Interface/Science/AreaIcons/portable_attractor.rsi/meta.json
@@ -1,14 +1,0 @@
-{
-  "version": 1,
-  "size": {
-    "x": 11,
-    "y": 11
-  },
-  "license": "CC-BY-SA-4.0",
-  "copyright": "Created by TheShuEd (Github)",
-  "states": [
-    {
-      "name": "icon"
-    }
-  ]
-}
